Add file needed for recent versions of cime
For versions of cime that use the newer cmake macros-based method of
porting instead of config_compilers.xml.

Note that I no longer need to duplicate the standard gnu section (as is
done in my config_compilers.xml): I had done that in order to remove
'-mcmodel=medium', but that isn't there anymore in CESM's standard gnu
CMake macros file.