Remove unneeded netcdf stuff in LDFLAGS
I have tested the CTSM pFUnit build without this and it works. I am
having problems with the CIME / CESM_share pFUnit build on my mac, so I
can't test that; it may be that this change further breaks that build,
and that I need to revisit this in the future.