Use gcc rather than clang for the serial C compiler
I don't regularly use the serial C compiler for system builds, though I
do use it for the unit test builds. It seems good to make this change so
that (I think) I'm using the same compiler for both mpi and serial
builds (since I think that mpicc currently points to gcc, not clang).

I confirmed that the CTSM unit test build still works with this change.
I'm having independent problems with the CIME / CESM_share unit test
build, so I can't test this. An mpi-serial system test builds
successfully with this change, though fails at the beginning of the run
with:

 *** The MPI_Comm_f2c() function was called before MPI_INIT was invoked.
 *** This is disallowed by the MPI standard.
 *** Your MPI job will now abort.

(I haven't tried an mpi-serial build with clang to see if I run into the
same problem with clang.)